Bugfix (#6)
* Bugfix: Correct use statment for Worksheet class

The use statement for the PhpOffice\PhpSpreadsheet\Worksheet\Worksheet was either missing a filename or a directory name.
This fix corrects that. This will fix https://github.com/MewesK/TwigSpreadsheetBundle/issues/2

* Bugfix: Start cells at first column

A COLUMN_DEFAULT of 0 made the first data field appear in column Z
This commit let the first data field appear in column A

* Bugfix: Correct use statements and Class

Cell has also gotten it's own namespace in PhpOffice\PhpSpreadsheet so that is added.
The stringFromColumnIndex method is no longer to be found in the class Cell, but in the class Coordinate. This is also fixed.